requirements.txt
python-dotenv>=1.0.0
pynetbox>=6.0.0
pyyaml>=6.0.0

